@@ -1,19 +1,19 @@
 module com.samderlust/sangtoolbox
 
-go 1.16
+go 1.18
 
 require (
-	github.com/bketelsen/crypt v0.0.4 // indirect
 	github.com/briandowns/spinner v1.19.0
+	github.com/pkg/errors v0.9.1
+	github.com/spf13/cobra v1.5.0
+	golang.org/x/term v0.0.0-20220919170432-7a66f970e087
+)
+
+require (
 	github.com/fatih/color v1.13.0 // indirect
 	github.com/inconshreveable/mousetrap v1.0.1 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
-	github.com/pkg/errors v0.9.1
-	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
-	github.com/smartystreets/goconvey v1.6.4 // indirect
-	github.com/spf13/cobra v1.5.0
-	github.com/spf13/viper v1.13.0 // indirect
-	golang.org/x/crypto v0.0.0-20220926161630-eccd6366d1be
+	github.com/mattn/go-isatty v0.0.16 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 	golang.org/x/sys v0.0.0-20220928140112-f11e5e49a4ec // indirect
-	golang.org/x/term v0.0.0-20220919170432-7a66f970e087
 )
