@@ -11,7 +11,7 @@ import (
 )
 
 // CreateDirsRecursive create folders and files follow provided template
-func CreateDirsRecursive(template interface{}, cwd string) error {
+func CreateDirsRecursive(template interface{}, pwd string) error {
 	var mapType map[string]interface{}
 	var arrType []interface{}
 
@@ -20,7 +20,7 @@ func CreateDirsRecursive(template interface{}, cwd string) error {
 	}
 
 	for key, val := range template.(map[string]interface{}) {
-		parentPath := filepath.Join(cwd, key)
+		parentPath := filepath.Join(pwd, key)
 		if err := os.MkdirAll(parentPath, 0777); err != nil {
 			fmt.Println(err)
 		}
@@ -38,6 +38,42 @@ func CreateDirsRecursive(template interface{}, cwd string) error {
 				}
 			} else {
 				if err := os.Mkdir(childPath, 0777); err != nil {
+					return nil
+				}
+			}
+		}
+	}
+	return nil
+}
+
+// CreateDirsRecursiveWithName create folders and files follow provided template
+func CreateDirsRecursiveWithName(template interface{}, flagMap map[string]string, pwd string) error {
+	var mapType map[string]interface{}
+	var arrType []interface{}
+
+	if reflect.TypeOf(template) != reflect.TypeOf(mapType) {
+		return errors.New("invalid template format")
+	}
+
+	for key, val := range template.(map[string]interface{}) {
+		parentPath := filepath.Join(pwd, key)
+		if err := os.MkdirAll(StringTemplateReplace(parentPath, flagMap), 0777); err != nil {
+			fmt.Println(err)
+		}
+
+		if reflect.TypeOf(val) != reflect.TypeOf(arrType) {
+			return errors.New("invalid format")
+		}
+
+		for _, v := range val.([]interface{}) {
+			childPath := filepath.Join(parentPath, v.(string))
+			if strings.Contains(childPath, ".") {
+				_, err := os.Create(StringTemplateReplace(childPath, flagMap))
+				if err != nil {
+					return nil
+				}
+			} else {
+				if err := os.Mkdir(StringTemplateReplace(childPath, flagMap), 0777); err != nil {
 					return nil
 				}
 			}
